+++ backend/app-module/src/main/java/com/onulmwhe/backend/config/SecurityConfig.java
@@ -20,7 +20,7 @@ public class SecurityConfig {
         http
             .csrf(csrf -> csrf.disable())
             .authorizeHttpRequests(auth -> auth
-                .requestMatchers("/actuator/health", "/error","/api/v1/routes/generate" ).permitAll()
+                .requestMatchers("/actuator/health", "/error", "/api/v1/routes/**").permitAll()
                 .anyRequest().authenticated()
             )
             .oauth2Login(oauth2 -> oauth2